Delete Polyfill
IE9 is already obsolete, so the first polyfill is not necessary. Because of the same reason, I cut the link to the "Polyfill of `setInterval` which allows passing arguments to the callback in `core-js`".
Second is not a problem since most browsers has `Function.prototype.bind()` method support, and arrow function is now available among modern browsers excluding IE. It makes sense that write the simpler method first.